+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+# [1.20.0] - 2022-07-05
+
+### Added
+
+- Adnexo - Ax-opto
+- Dragino - LHT65
+- Terabee - Level Monitoring XL
+
 # [1.19.0] - 2022-06-30
 
 ### Added